DO NOT MERGE: Remove test that depends on tz data
Commenting out a test that will pass with <= 2017b
time zone data but fail on >= 2017c time zone data.

The has been removed for historic CTS releases because
there should be adequate coverage from the other tests
and it would be more complicated to provide a test
that works on a range of devices. We can expect OEMs
to release system images for releases >= O with 2017c
data. The affected test does not exist prior to O.

Bug: 68878031
Test: build only
Change-Id: I4320d6145a9d6506b5f427069a4fe0906158b070