Fix TestFormatter test for 2020a update [O-MR1]
This is a cherry-pick of the change below to make the
CtsLibcoreOjTestCases pass on the dev branch.

Bug: 155407785

====

Revert accidental change in test.

I accidentally reverted a change in TestFormatter to TestFormatter.java
in commit If50d927265136e8de964b54ae879291c7be85cfc.

Since the original change was unrelated to DecimalFormat, reverting it
was not supposed to happen in that commit.

The issue was not caught because CtsLibcoreOjTestCases wasn't run for
that test (only CtsLibcoreTestCases).

Bug: 69418703
Test: CtsLibcoreOjTestCases
Change-Id: If6a2dbb615a70739350cbae91cc33754afe2b793
(cherry picked from commit 044bd5e685b4bd9302ace9f77b7339cbe7e6f893)